# docs/developer-docs/public/indexer/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tari_indexer",
    "description": "The Tari indexer implementation",
    "contact": {
      "name": "The Tari Development Community"
    },
    "license": {
      "name": "BSD-3-Clause",
      "identifier": "BSD-3-Clause"
    },
    "version": "0.39.2"
  },
  "paths": {
...
    "/epoch-manager/stats": {
      "get": {
        "tags": [
          "handlers::misc"
        ],
        "description": "Get epoch manager stats",
        "operationId": "get_epoch_manager_stats",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetEpochManagerStatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/identity": {
      "get": {
        "tags": [
          "handlers::misc"
        ],
        "description": "Get indexer network identity information",
        "operationId": "get_identity",
        "responses": {
          "200": {
            "description": "Indexer network identity info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetIdentityResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get identity info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "GetEpochManagerStatsResponse": {
        "type": "object",
        "required": [
          "current_epoch",
          "current_block_height",
          "current_block_hash"
        ],
        "properties": {
          "current_block_hash": {
            "type": "string"
          },
          "current_block_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "current_epoch": {
            "type": "integer",
            "format": "int64",
            "description": "The current epoch according to the indexer's epoch oracle view",
            "minimum": 0
          }
        }
      },
      "GetIdentityResponse": {
        "type": "object",
        "required": [
          "peer_id",
          "public_key",
          "public_addresses"
        ],
        "properties": {
          "peer_id": {
            "type": "string"
          },
          "public_addresses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "public_key": {
            "type": "string"
          }
        }
      },
...
    }
  }
}